+++ python/pathway/internals/universe_solver.py
@@ -1,11 +1,13 @@
 # Copyright © 2024 Pathway
 
 import itertools
+import warnings
 from collections import defaultdict
 from collections.abc import Iterator
 
 from pysat.solvers import Solver
 
+from pathway.internals.trace import Trace, _format_frame
 from pathway.internals.universe import Universe
 
 
@@ -13,17 +15,27 @@ class UniverseSolver:
     universe_vars: dict[Universe, int]
     var_counter: Iterator[int]
     solver: Solver
+    no_warn: set[Universe]
 
     def __init__(self):
         self.solver = Solver(name="g4")
         self.var_counter = itertools.count(start=1)
         self.universe_vars = defaultdict(lambda: next(self.var_counter))
+        self.no_warn = set()
 
     def register_as_equal(self, left: Universe, right: Universe) -> None:
-        self.register_as_subset(left, right)
-        self.register_as_subset(right, left)
+        self._register_as_equal(left, right)
+        self._validate_nonempty_universes()
+
+    def _register_as_equal(self, left: Universe, right: Universe) -> None:
+        self._register_as_subset(left, right)
+        self._register_as_subset(right, left)
 
     def register_as_subset(self, subset: Universe, superset: Universe) -> None:
+        self._register_as_subset(subset, superset)
+        self._validate_nonempty_universes()
+
+    def _register_as_subset(self, subset: Universe, superset: Universe) -> None:
         varA = self.universe_vars[subset]
         varB = self.universe_vars[superset]
         # varA => varB
@@ -42,9 +54,15 @@ class UniverseSolver:
     def register_as_difference(
         self, result: Universe, setLeft: Universe, setRight: Universe
     ) -> None:
+        self._register_as_difference(result, setLeft, setRight)
+        self._validate_nonempty_universes()
+
+    def _register_as_difference(
+        self, result: Universe, setLeft: Universe, setRight: Universe
+    ) -> None:
         """result = setLeft - setRight"""
-        self.register_as_subset(result, setLeft)
-        self.register_as_disjoint(result, setRight)
+        self._register_as_subset(result, setLeft)
+        self._register_as_disjoint(result, setRight)
         varResult = self.universe_vars[result]
         varLeft = self.universe_vars[setLeft]
         varRight = self.universe_vars[setRight]
@@ -57,8 +75,12 @@ class UniverseSolver:
         return result
 
     def register_as_intersection(self, result: Universe, *args: Universe) -> None:
+        self._register_as_intersection(result, *args)
+        self._validate_nonempty_universes()
+
+    def _register_as_intersection(self, result: Universe, *args: Universe) -> None:
         for arg in args:
-            self.register_as_subset(result, arg)
+            self._register_as_subset(result, arg)
 
         result_var = self.universe_vars[result]
         args_var = [self.universe_vars[arg] for arg in args]
@@ -71,8 +93,12 @@ class UniverseSolver:
         return result
 
     def register_as_union(self, result: Universe, *args: Universe) -> None:
+        self._register_as_union(result, *args)
+        self._validate_nonempty_universes()
+
+    def _register_as_union(self, result: Universe, *args: Universe) -> None:
         for arg in args:
-            self.register_as_subset(arg, result)
+            self._register_as_subset(arg, result)
 
         result_var = self.universe_vars[result]
         args_var = [self.universe_vars[arg] for arg in args]
@@ -103,6 +129,11 @@ class UniverseSolver:
         return True
 
     def register_as_disjoint(self, *args: Universe) -> None:
+        self._register_as_disjoint(*args)
+        self._validate_nonempty_universes()
+
+    def _register_as_disjoint(self, *args: Universe) -> None:
+        # TODO: this code might be doable with O(n) checks, not O(n^2)
         vars = [self.universe_vars[arg] for arg in args]
         for i in range(len(vars)):
             for j in range(i):
@@ -111,10 +142,37 @@ class UniverseSolver:
 
     def query_is_empty(self, setA: Universe) -> bool:
         varA = self.universe_vars[setA]
-        if self.solver.solve(assumptions=[varA]):
-            return False
-        return True
+        return not self.solver.solve(assumptions=[varA])
 
-    def register_as_empty(self, setA: Universe) -> None:
+    def register_as_empty(self, setA: Universe, no_warn: bool = True) -> None:
+        self._register_as_empty(setA)
+        if no_warn:
+            self.no_warn.add(setA)
+        self._validate_nonempty_universes()
+
+    def _register_as_empty(self, setA: Universe) -> None:
         varA = self.universe_vars[setA]
         self.solver.add_clause([-varA])
+
+    def _validate_nonempty_universes(self) -> None:
+        for univ in self.universe_vars.keys():
+            if univ in self.no_warn:
+                continue
+            if self.query_is_empty(univ):
+                try:
+                    trace = univ.lineage.trace
+                except AttributeError:
+                    trace = Trace.from_traceback()
+                frame = trace.user_frame
+                if frame is None:
+                    warnings.warn(
+                        "Found universe that is always empty, but wasn't declared as such --"
+                        + " this is potentially a bug.\n"
+                    )
+                else:
+                    warnings.warn(
+                        "Found universe that is always empty, but wasn't declared as such --"
+                        + " this is potentially a bug.\n"
+                        + _format_frame(frame)
+                    )
+                self.no_warn.add(univ)