+++ CHANGELOG.md
@@ -5,6 +5,9 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ## [Unreleased]
 
+### Added
+- Pathway now warns when unintentionally creating Table with empty universe.
+
 ## [0.9.0] - 2024-04-18
 
 ### Added